refactor: 행정구역 코드 타입 변경

+++ src/main/java/io/dnd/goyo/domain/place/entity/Place.java
@@ -92,7 +92,7 @@ public class Place extends BaseEntity {
             String name,
             PlaceCategory category,
             Point location,
-            Integer regionCode,
+            Long regionCode,
             String addressDetail,
             User user,
             Integer floorInfo,
